preferences: Refactor to be more self-contained
Create the preferences window on demand instead of at startup; destroy
instead of hiding when closing. Instead of manipulating its widgets
directly from other files, use signal handlers to make it more
self-contained.
