+++ src/app.h
@@ -17,7 +17,6 @@
 #include <gtksourceview/gtksource.h>
 
 #include "document.h"
-#include "prefs.h"
 
 #define I7_TYPE_APP            (i7_app_get_type())
 #define I7_APP(obj)            (G_TYPE_CHECK_INSTANCE_CAST((obj), I7_TYPE_APP, I7App))
@@ -51,8 +50,6 @@ typedef struct {
 typedef struct {
 	GtkApplication parent_instance;
 
-	/* Public preferences dialog */
-	I7PrefsWindow *prefs;
 	/* Already-compiled regices */
 	GRegex *regices[I7_APP_NUM_REGICES];
 } I7App;
@@ -118,7 +115,6 @@ void i7_app_set_print_settings(I7App *self, GtkPrintSettings *settings);
 
 GtkSourceStyleSchemeManager *i7_app_get_color_scheme_manager(I7App *self);
 
-void i7_app_present_prefs_window(I7App *self);
 void i7_app_update_css(I7App *self);
 
 char *i7_app_get_font_family(I7App *self);
