+++ src/ui/prefs.ui
@@ -37,7 +37,6 @@ The Customs Wharf is a room. [change the description if the cask is open] "Amid 
     <property name="icon-name">com.inform7.IDE</property>
     <property name="type-hint">dialog</property>
     <signal name="close" handler="gtk_widget_hide" swapped="no"/>
-    <signal name="delete-event" handler="gtk_widget_hide_on_delete" swapped="no"/>
     <child internal-child="vbox">
       <object class="GtkBox" id="dialog-vbox2">
         <property name="visible">True</property>
@@ -56,7 +55,7 @@ The Customs Wharf is a room. [change the description if the cask is open] "Amid 
                 <property name="can-focus">True</property>
                 <property name="receives-default">True</property>
                 <property name="use-underline">True</property>
-                <signal name="clicked" handler="gtk_widget_hide" object="I7PrefsWindow" swapped="yes"/>
+                <signal name="clicked" handler="gtk_widget_destroy" object="I7PrefsWindow" swapped="yes"/>
               </object>
               <packing>
                 <property name="expand">False</property>
